Fix for refresh issue
Little bit hacky, but it’ll do for right now.

On page load, angular will run the `checkSession` function, which hits
`router.php/auth`. If there is a `result[0]`, the user isn’t
authenticated. The `0` object is only returned when the user isn’t
authenticated (it contains a variable `text` which equals ‘User is not
logged in’ string) - we should look at a more concrete way of doing
this maybe?

This whole piece `SessionCtrl` piece could probably do with being
rewritten since we know more about Angular now.